+++ src/casl/checkPolicies.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { PolicyHandler } from './policy.handler';
+import { PolicyHandler } from './handlers/policy.handler';
 
 export const CHECK_POLICIES_KEY = 'check_policy';
 export const CheckPolicies = (...handlers: PolicyHandler[]) =>
